requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.6.4

# Vector database (FAISS)
faiss-cpu==1.7.4
numpy==1.24.3

# LLM and embeddings
openai>=1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Document processing
pdfplumber==0.11.0
python-docx==1.1.0
requests==2.31.0

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1
